fixed trying to input string numbers for the abundance table

+++ bpaotu/bpaotu/importer.py
@@ -404,7 +404,7 @@ class DataImporter:
                         bpa_id = column
                         count = row[column]
                         try:
-                            count = int(count)
+                            count = float(count)
                         except:
                             logger.warning('count invalid, defaulting to 0')
                             count = 0
